# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Pothos API
  description: Self-hostable budget and expense tracking application API.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pth_<64-char-hex>
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          description: Additional validation details
...
    Budget:
      type: object
      required: [id, userId, categoryId, amount, month, year, isRecurring, isCommitted, spent, remaining, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        categoryId:
          type: string
        amount:
          type: integer
          description: Budget limit in minor units (always positive)
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        isRecurring:
          type: boolean
        isCommitted:
          type: boolean
          description: When true, this budget is counted as committed spending in the overview report
        spent:
          type: integer
          description: Calculated from expense transactions in the month (positive)
        remaining:
          type: integer
          description: amount - spent
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    ParsedTransaction:
      type: object
      required: [id, userId, type, amount, date, description, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        pendingMessageId:
          type: [string, "null"]
        accountId:
          type: [string, "null"]
        categoryId:
          type: [string, "null"]
        type:
          type: string
          enum: [income, expense]
        amount:
          type: integer
          description: Positive integer. Sign applied when approved.
        date:
          type: integer
        description:
          type: string
        notes:
          type: [string, "null"]
        status:
          type: string
          enum: [pending_review, approved, rejected]
        accountName:
          type: [string, "null"]
        categoryName:
          type: [string, "null"]
        emailSubject:
          type: [string, "null"]
        createdAt:
          type: integer
        updatedAt:
          type: integer
    PendingMessage:
      type: object
      required: [id, userId, rawContent, source, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        rawContent:
          type: string
        subject:
          type: [string, "null"]
        source:
          type: string
          enum: [imap]
        status:
          type: string
          enum: [pending, processed, failed]
        error:
          type: [string, "null"]
        createdAt:
          type: integer
        updatedAt:
          type: integer
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /budgets:
    get:
      summary: List budgets with spent/remaining for a month
      tags: [Budgets]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Defaults to current month
        - name: year
          in: query
          schema:
            type: integer
          description: Defaults to current year
      responses:
        "200":
          description: Budgets with calculated spent and remaining amounts. Auto-generates recurring budgets for the current month on first view.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"

    post:
      summary: Create or update (upsert) a budget
      tags: [Budgets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, amount, month, year]
              properties:
                categoryId:
                  type: string
                amount:
                  type: integer
                  minimum: 1
                  maximum: 1000000000
                month:
                  type: integer
                  minimum: 1
                  maximum: 12
                year:
                  type: integer
                  minimum: 2000
                  maximum: 2100
                isRecurring:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "200":
          description: Updated existing budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    delete:
      summary: Delete budget
      tags: [Budgets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Reports ───────────────────────────────────────────────────────────────
...
  /parse-queue:
    get:
      summary: List raw pending emails for local LLM processing
      tags: [ParseQueue]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processed, failed]
            default: pending
      responses:
        "200":
          description: Pending messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingMessage"
  /parse-queue/{id}/submit:
    post:
      summary: Submit a locally parsed result to the review queue
      tags: [ParseQueue]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: pending_message id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, amount, date, description]
              properties:
                type:
                  type: string
                  enum: [income, expense]
                amount:
                  type: integer
                  minimum: 1
                date:
                  type: integer
                description:
                  type: string
                  minLength: 1
                  maxLength: 200
                accountId:
                  type: [string, "null"]
                categoryId:
                  type: [string, "null"]
                notes:
                  type: [string, "null"]
                bypassReview:
                  type: boolean
                  default: false
                  description: When true, creates the transaction directly instead of adding to the review queue
      responses:
        "201":
          description: Parsed transaction created in review queue (or actual transaction if bypassReview=true)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedTransaction"
        "404":
          description: Pending message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Message already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /parse-queue/{id}/dismiss:
    post:
      summary: Dismiss a pending message (marks as failed)
      tags: [ParseQueue]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dismissed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── API Keys ──────────────────────────────────────────────────────────────
...